Rework tests to unify development and testing
Summary:
The previous tests used beaker to create a VM and run the manifests, which is
silly, since quite separately there is a Vagrantfile to create a development
environment in which to run the manifests.

Now, the tests use the same Vagrantfile to create a VM, then run the tests in
them, and make assertions with serverspec directly.

This serves as a proof of concept for our new CI strategey. Ref T4.

Test Plan: Tests run on my workstation. Just "rake spec" does the job. One fails (testing idempotency), but this is an issue that already existed, but wasn't tested.

Reviewers: #macpros_internal

Maniphest Tasks: T4

Differential Revision: http://phabricator.macprofessionals.lan/D5